[feature] JAX-RS client Conjure generator
**Why**
Many use cases are better suited for blocking clients. There is no
doubt that asynchronous clients can be more powerful, however
they're often both unnecessary and dangerous. Threads are hard,
and it's best to avoid thinking about them until we need to.

The existing Jersey generator is sufficient in many cases,
but it's entirely broken for streaming binary data. Because
the Jersey generator outputs code for both clients and servers,
it must be a jack of all trades and master of none. For example,
there's no reason for server interfaces to scramble argument
order or provide default methods, but they do so because it's
helpful when the interface is reused by clients. There's also
no reason for client interfaces to return a server-side
`StreamingOutput` from binary methods because the type is
unhelpful to clients. By decoupling client and server
interfaces we allow ourselves to generate more accurate
code for both.

**What**
This change adds a `--jaxRsClient` flag to the CLI to generate
client-specific interfaces. Our JaxRsClient interfaces differ
in a few ways:
1. Generated interfaces names are suffixed with `JaxRsClient`
   to differentiate them from Jersey services and avoid classpath
   conflicts.
2. `returns: binary` results in methods which return InputStream
3. `returns: optional<binary>` results in methods which return
   `Optional<InputStream>`
4. The generator ignores `--requireNotNullAuthAndBodyParams`
   because feign doesn't use doesn't use
   `javax.validation.constraints`

**Future Work**
I plan to update `conjure-java-runtime` JaxRsClients to avoid fully
buffering binary request and response bodies on heap. This is not
a blocker for the generator because we have decided that buffering
binary request bodies is acceptable, so we can temporarily allow
the same behavior for responses.

+++ conjure-java-core/src/test/java/com/palantir/conjure/java/TestBase.java
@@ -53,7 +53,12 @@ public abstract class TestBase {
         for (Path file : files) {
             Path output = outputDir.resolve(file.getFileName() + suffix);
             if (Boolean.valueOf(System.getProperty("recreate", "false"))) {
-                Files.delete(output);
+                if (Files.exists(output)) {
+                    Files.delete(output);
+                }
+                if (!Files.isDirectory(output.getParent())) {
+                    Files.createDirectories(output.getParent());
+                }
                 Files.copy(file, output);
             }
             assertThat(readFromFile(file)).isEqualTo(readFromFile(output));
